Move Event class into own file
